Move the code related to scanDirectories and scanFiles into StorageDB
